PCI-322: Push business logic down into a service layer.
