+++ src/main/java/uk/gov/companieshouse/items/orders/api/ItemType.java
@@ -1,6 +1,6 @@
 package uk.gov.companieshouse.items.orders.api;
 
-import uk.gov.companieshouse.items.orders.api.dto.ItemDTO;
+import uk.gov.companieshouse.items.orders.api.model.Item;
 import uk.gov.companieshouse.items.orders.api.model.ItemCosts;
 
 /**
@@ -12,8 +12,8 @@ public enum ItemType {
     CERTIFIED_COPY("certified-copy", "certified-copy"),
     SCAN_ON_DEMAND("scan-on-demand", "scan-on-demand") {
         @Override
-        public void populatePostalDelivery(final ItemDTO itemDTO) {
-            itemDTO.setPostalDelivery(false);
+        public void populatePostalDelivery(final Item item) {
+            item.setPostalDelivery(false);
         }
     };
 
@@ -27,29 +27,29 @@ public enum ItemType {
 
     /**
      * Populates the read only fields of the item DTO provided.
-     * @param itemDTO the DTO with read only fields
+     * @param item the item with read only fields
      */
-    public void populateReadOnlyFields(final ItemDTO itemDTO) {
-        populateDescriptionFields(itemDTO);
-        populateItemCosts(itemDTO);
-        populatePostalDelivery(itemDTO);
+    public void populateReadOnlyFields(final Item item) {
+        populateDescriptionFields(item);
+        populateItemCosts(item);
+        populatePostalDelivery(item);
     }
 
     /**
      * Populates the description fields to facilitate UI text rendering.
-     * @param itemDTO the DTO bearing text for UI rendering
+     * @param item the item bearing text for UI rendering
      */
-    void populateDescriptionFields(final ItemDTO itemDTO) {
-        itemDTO.setDescriptionIdentifier(itemType);
-        itemDTO.setKind(kind);
+    void populateDescriptionFields(final Item item) {
+        item.setDescriptionIdentifier(itemType);
+        item.setKind(kind);
         // TODO PCI-505 Populate description and description values when we know what these are.
     }
 
     /**
      * Populates the item costs fields.
-     * @param itemDTO the DTO bearing the item costs
+     * @param item the item bearing the item costs
      */
-    void populateItemCosts(final ItemDTO itemDTO) {
+    void populateItemCosts(final Item item) {
         final ItemCosts costs = new ItemCosts();
 
         // TODO PCI-506 Retrieve the actual costs as appropriate.
@@ -58,14 +58,14 @@ public enum ItemType {
         costs.setPostageCost("3");
         costs.setTotalCost("4");
 
-        itemDTO.setItemCosts(costs);
+        item.setItemCosts(costs);
     }
 
     /**
      * Populates the postal delivery field as appropriate for this type.
-     * @param itemDTO the DTO
+     * @param item the item
      */
-    void populatePostalDelivery(final ItemDTO itemDTO) {
-        itemDTO.setPostalDelivery(true);
+    void populatePostalDelivery(final Item item) {
+        item.setPostalDelivery(true);
     }
 }
