+++ src/main/java/uk/gov/companieshouse/items/orders/api/controller/CertificateItemsController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 import uk.gov.companieshouse.items.orders.api.dto.CertificateItemDTO;
 import uk.gov.companieshouse.items.orders.api.mapper.CertificateItemMapper;
 import uk.gov.companieshouse.items.orders.api.model.CertificateItem;
+import uk.gov.companieshouse.items.orders.api.service.CertificateItemService;
 import uk.gov.companieshouse.items.orders.api.validator.CreateItemRequestValidator;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
@@ -15,7 +16,6 @@ import uk.gov.companieshouse.logging.LoggerFactory;
 import javax.validation.Valid;
 import java.util.List;
 
-import static uk.gov.companieshouse.items.orders.api.ItemType.CERTIFICATE;
 import static uk.gov.companieshouse.items.orders.api.ItemsApiApplication.APPLICATION_NAMESPACE;
 
 @RestController
@@ -25,16 +25,19 @@ public class CertificateItemsController {
 
     private final CreateItemRequestValidator validator;
     private final CertificateItemMapper mapper;
+    private final CertificateItemService service;
 
-    public CertificateItemsController(final CreateItemRequestValidator validator, final CertificateItemMapper mapper) {
+    public CertificateItemsController(final CreateItemRequestValidator validator,
+                                      final CertificateItemMapper mapper,
+                                      final CertificateItemService service) {
         this.validator = validator;
         this.mapper = mapper;
+        this.service = service;
     }
 
     @PostMapping("${uk.gov.companieshouse.items.orders.api.path}")
     public ResponseEntity<Object> createCertificateItem(final @Valid @RequestBody CertificateItemDTO certificateItemDTO)
     {
-
         LOGGER.info("ENTERING createCertificateItem(" + certificateItemDTO + ")");
 
         final List<String> errors = validator.getValidationErrors(certificateItemDTO);
@@ -42,18 +45,12 @@ public class CertificateItemsController {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ApiError(HttpStatus.BAD_REQUEST, errors));
         }
 
-        CERTIFICATE.populateReadOnlyFields(certificateItemDTO);
+        CertificateItem item = mapper.certificateItemDTOtoCertificateItem(certificateItemDTO);
+        item = service.createCertificateItem(item);
+        final CertificateItemDTO createdCertificateItemDTO = mapper.certificateItemToCertificateItemDTO(item);
 
-        // TODO PCI-324 ID will be generated as per
-        //  https://companieshouse.atlassian.net/wiki/spaces/DEV/pages/1258094916/Certificates+API+High+Level+Development+Design#CertificatesAPIHighLevelDevelopmentDesign-Traceability
-        certificateItemDTO.setId("CHS1");
-
-        final CertificateItem item = mapper.certificateItemDTOtoCertificateItem(certificateItemDTO);
-        // TODO Remove this?
-        LOGGER.info("Mapped Certificate item = " + item);
-
-        LOGGER.info("EXITING createCertificateItem() with " + certificateItemDTO);
-        return ResponseEntity.status(HttpStatus.CREATED).body(certificateItemDTO);
+        LOGGER.info("EXITING createCertificateItem() with " + createdCertificateItemDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdCertificateItemDTO);
     }
 
 }
