+++ src/main/java/uk/gov/companieshouse/items/orders/api/model/Item.java
@@ -1,0 +1,116 @@
+package uk.gov.companieshouse.items.orders.api.model;
+
+import com.google.gson.Gson;
+import org.springframework.data.annotation.Id;
+
+import java.util.Map;
+
+/**
+ * An instance of this represents an item of any type.
+ */
+public class Item {
+
+    @Id
+    private String id;
+
+    private String companyNumber;
+
+    private String description;
+
+    private String descriptionIdentifier;
+
+    private Map<String, String> descriptionValues;
+
+    private ItemCosts itemCosts;
+
+    private CertificateItemOptions itemOptions;
+
+    private String kind;
+
+    private boolean isPostalDelivery;
+
+    private int quantity;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getCompanyNumber() {
+        return companyNumber;
+    }
+
+    public void setCompanyNumber(String companyNumber) {
+        this.companyNumber = companyNumber;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getDescriptionIdentifier() {
+        return descriptionIdentifier;
+    }
+
+    public void setDescriptionIdentifier(String descriptionIdentifier) {
+        this.descriptionIdentifier = descriptionIdentifier;
+    }
+
+    public Map<String, String> getDescriptionValues() {
+        return descriptionValues;
+    }
+
+    public void setDescriptionValues(Map<String, String> descriptionValues) {
+        this.descriptionValues = descriptionValues;
+    }
+
+    public ItemCosts getItemCosts() {
+        return itemCosts;
+    }
+
+    public void setItemCosts(ItemCosts itemCosts) {
+        this.itemCosts = itemCosts;
+    }
+
+    public CertificateItemOptions getItemOptions() {
+        return itemOptions;
+    }
+
+    public void setItemOptions(CertificateItemOptions itemOptions) {
+        this.itemOptions = itemOptions;
+    }
+
+    public String getKind() {
+        return kind;
+    }
+
+    public void setKind(String kind) {
+        this.kind = kind;
+    }
+
+    public boolean isPostalDelivery() {
+        return isPostalDelivery;
+    }
+
+    public void setPostalDelivery(boolean postalDelivery) {
+        isPostalDelivery = postalDelivery;
+    }
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+
+    @Override
+    public String toString() { return new Gson().toJson(this); }
+}
